refactor(docs): share reloadable code logic
