+++ .gitignore
@@ -87,7 +87,8 @@ versions.json
 /docs/api/api-search-index.json
 /docs/public/api-diff-index.json
 /docs/public/faker.js
-/docs/locales/*.md
+/docs/locales/*
+!/docs/locales/.gitkeep
 
 # Faker
 TAGS
