+++ docs/locales/.gitkeep
@@ -1,0 +1,1 @@
+Everything in this folder is auto generated.
